--- src/plugins/image-paste-handler/index.js
@@ -1,40 +1,0 @@
-import decorateComponentWithProps from 'decorate-component-with-props'
-import ImageBlock from './image-block'
-
-/**
- * image paste handler
- */
-export default (config = {}) => {
-  let ImageComponent = ImageBlock
-  if (config.decorator) {
-    ImageComponent = config.decorator(ImageBlock)
-  }
-  const theme = config.theme || {}
-
-  const ThemedImage = decorateComponentWithProps(ImageComponent, { theme })
-
-  return {
-    // handle single image paste and insert image by tool click
-    blockRendererFn: (block, { getEditorState }) => {
-      if (block.getType() === 'atomic') {
-        const key = block.getEntityAt(0)
-        if (!key) {
-          return null
-        }
-        const contentState = getEditorState().getCurrentContent()
-        const entity = contentState.getEntity(key)
-        if (!entity) return null
-
-        const type = entity.getType().toLowerCase()
-        if (type === 'image') {
-          return {
-            component: ThemedImage,
-            editable: false
-          }
-        }
-
-        return null
-      }
-    }
-  }
-}
